Tweak cache usage

+++ DonutEngine/src/Donut/ECS/ECSComponent.cpp
@@ -5,7 +5,7 @@ namespace Donut {
 
 	uint32_t BaseECSComponent::RegisterComponent(ECSComponentCreateFunction createfn, ECSComponentFreeFunction freefn, size_t size)
 	{
-		uint32_t index = m_ComponentTypes.size();
+		uint32_t index = (uint32_t)m_ComponentTypes.size();
 		m_ComponentTypes.push_back({ createfn, freefn, size });
 		return index;
 	}
